requirements.txt
langchain
faiss-cpu
openai
tiktoken
streamlit
pytesseract
pdf2image
